moved EmailTableModel and deleted unecessary files
