--- src/main/java/interface_adapter/view_dashboard/DashboardViewModel.java
@@ -1,13 +1,0 @@
-package interface_adapter.view_dashboard;
-
-import interface_adapter.ViewModel;
-
-/**
- * The View Model for the Dashboard View.
- */
-public class DashboardViewModel extends ViewModel<DashboardState>{
-    public DashboardViewModel() {
-        super("viewing dashboard");
-        setState(new DashboardState());
-    }
-}
